requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-CORS==4.0.0 